refactor(transcripts): extract pure WAV chunking from TranscriptionService
The service mixed ~230 lines of pure audio work — chunk sizing, m4a->WAV
decode/re-encode, and WAV header/PCM byte writing — into a class otherwise
concerned with the queue, OpenAI client, diarization routing, and saving.
Move it to a standalone src/services/audioChunker.ts of pure functions
(createChunkFiles/getMimeType/...), shrinking the class 631->400 LOC.

The chunker tests previously kept hand-maintained COPIES of getMimeType,
shouldConvertToWav, createBinaryChunkFiles, and writeWavHeader, and reached
createChunkFiles via an `as unknown as` cast on a live instance. They now
live in audioChunker.test.ts and exercise the real exported functions, so
they can no longer silently drift from production. No behavior change.
