+++ src/services/TranscriptionService.test.ts
@@ -83,51 +83,10 @@ describe("TranscriptionService", () => {
 		});
 	});
 
-	describe("getMimeType", () => {
-		test("returns correct mime types for audio formats", () => {
-			const getMimeType = (fileExtension: string): string => {
-				switch (fileExtension.toLowerCase()) {
-					case "mp3":
-						return "audio/mp3";
-					case "m4a":
-						return "audio/mp4";
-					case "ogg":
-						return "audio/ogg";
-					case "wav":
-						return "audio/wav";
-					case "flac":
-						return "audio/flac";
-					case "webm":
-						return "audio/webm";
-					default:
-						return "audio/mpeg";
-				}
-			};
-
-			expect(getMimeType("mp3")).toBe("audio/mp3");
-			expect(getMimeType("MP3")).toBe("audio/mp3");
-			expect(getMimeType("m4a")).toBe("audio/mp4");
-			expect(getMimeType("ogg")).toBe("audio/ogg");
-			expect(getMimeType("wav")).toBe("audio/wav");
-			expect(getMimeType("flac")).toBe("audio/flac");
-			expect(getMimeType("webm")).toBe("audio/webm");
-			expect(getMimeType("unknown")).toBe("audio/mpeg");
-		});
-	});
-
-	describe("shouldConvertToWav", () => {
-		test("returns true for m4a files", () => {
-			const shouldConvertToWav = (extension: string, mimeType: string): boolean => {
-				const normalizedExtension = extension.toLowerCase();
-				return normalizedExtension === "m4a" || mimeType === "audio/mp4";
-			};
-
-			expect(shouldConvertToWav("m4a", "audio/mp4")).toBe(true);
-			expect(shouldConvertToWav("M4A", "audio/mp4")).toBe(true);
-			expect(shouldConvertToWav("mp3", "audio/mp4")).toBe(true);
-			expect(shouldConvertToWav("mp3", "audio/mpeg")).toBe(false);
-		});
-	});
+	// NOTE: the audio chunking + WAV encoding tests (getMimeType, shouldConvertToWav,
+	// createBinaryChunkFiles, writeWavHeader, createChunkFiles) moved to
+	// audioChunker.test.ts, where they exercise the real exported functions instead
+	// of inline copies. This file keeps the service-orchestration tests.
 
 	describe("getEpisodeKey", () => {
 		test("generates unique key from podcast name and title", () => {
@@ -136,146 +95,6 @@ describe("TranscriptionService", () => {
 			};
 
 			expect(getEpisodeKey(mockEpisode)).toBe("Test Podcast:Test Episode");
-		});
-	});
-
-	describe("createBinaryChunkFiles", () => {
-		const CHUNK_SIZE_BYTES = 20 * 1024 * 1024;
-
-		function createBinaryChunkFiles(
-			buffer: ArrayBuffer,
-			basename: string,
-			extension: string,
-			mimeType: string,
-		): File[] {
-			if (buffer.byteLength <= CHUNK_SIZE_BYTES) {
-				return [
-					new File([buffer], `${basename}.${extension}`, {
-						type: mimeType,
-					}),
-				];
-			}
-
-			const files: File[] = [];
-			for (
-				let offset = 0, index = 0;
-				offset < buffer.byteLength;
-				offset += CHUNK_SIZE_BYTES, index++
-			) {
-				const chunk = buffer.slice(offset, offset + CHUNK_SIZE_BYTES);
-				files.push(
-					new File([chunk], `${basename}.part${index}.${extension}`, {
-						type: mimeType,
-					}),
-				);
-			}
-
-			return files;
-		}
-
-		test("returns single file when buffer is smaller than chunk size", () => {
-			const smallBuffer = new ArrayBuffer(1024);
-			const files = createBinaryChunkFiles(smallBuffer, "test", "mp3", "audio/mpeg");
-
-			expect(files).toHaveLength(1);
-			expect(files[0].name).toBe("test.mp3");
-			expect(files[0].type).toBe("audio/mpeg");
-			expect(files[0].size).toBe(1024);
-		});
-
-		test("returns multiple files when buffer exceeds chunk size", () => {
-			const largeBuffer = new ArrayBuffer(CHUNK_SIZE_BYTES * 2 + 1024);
-			const files = createBinaryChunkFiles(largeBuffer, "test", "mp3", "audio/mpeg");
-
-			expect(files).toHaveLength(3);
-			expect(files[0].name).toBe("test.part0.mp3");
-			expect(files[1].name).toBe("test.part1.mp3");
-			expect(files[2].name).toBe("test.part2.mp3");
-		});
-	});
-
-	describe("writeWavHeader", () => {
-		const WAV_HEADER_SIZE = 44;
-		const PCM_BYTES_PER_SAMPLE = 2;
-
-		function writeString(view: DataView, offset: number, str: string): void {
-			for (let i = 0; i < str.length; i++) {
-				view.setUint8(offset + i, str.charCodeAt(i));
-			}
-		}
-
-		function writeWavHeader(
-			view: DataView,
-			sampleRate: number,
-			numChannels: number,
-			sampleCount: number,
-		): void {
-			const blockAlign = numChannels * PCM_BYTES_PER_SAMPLE;
-			const byteRate = sampleRate * blockAlign;
-			const dataSize = sampleCount * blockAlign;
-			writeString(view, 0, "RIFF");
-			view.setUint32(4, 36 + dataSize, true);
-			writeString(view, 8, "WAVE");
-			writeString(view, 12, "fmt ");
-			view.setUint32(16, 16, true);
-			view.setUint16(20, 1, true);
-			view.setUint16(22, numChannels, true);
-			view.setUint32(24, sampleRate, true);
-			view.setUint32(28, byteRate, true);
-			view.setUint16(32, blockAlign, true);
-			view.setUint16(34, PCM_BYTES_PER_SAMPLE * 8, true);
-			writeString(view, 36, "data");
-			view.setUint32(40, dataSize, true);
-		}
-
-		test("writes correct RIFF header", () => {
-			const buffer = new ArrayBuffer(WAV_HEADER_SIZE);
-			const view = new DataView(buffer);
-
-			writeWavHeader(view, 44100, 2, 44100);
-
-			const riff = String.fromCharCode(
-				view.getUint8(0),
-				view.getUint8(1),
-				view.getUint8(2),
-				view.getUint8(3),
-			);
-			expect(riff).toBe("RIFF");
-
-			const wave = String.fromCharCode(
-				view.getUint8(8),
-				view.getUint8(9),
-				view.getUint8(10),
-				view.getUint8(11),
-			);
-			expect(wave).toBe("WAVE");
-
-			const fmt = String.fromCharCode(
-				view.getUint8(12),
-				view.getUint8(13),
-				view.getUint8(14),
-				view.getUint8(15),
-			);
-			expect(fmt).toBe("fmt ");
-
-			const data = String.fromCharCode(
-				view.getUint8(36),
-				view.getUint8(37),
-				view.getUint8(38),
-				view.getUint8(39),
-			);
-			expect(data).toBe("data");
-		});
-
-		test("writes correct sample rate and channels", () => {
-			const buffer = new ArrayBuffer(WAV_HEADER_SIZE);
-			const view = new DataView(buffer);
-
-			writeWavHeader(view, 44100, 2, 1000);
-
-			expect(view.getUint16(22, true)).toBe(2);
-			expect(view.getUint32(24, true)).toBe(44100);
-			expect(view.getUint16(34, true)).toBe(16);
 		});
 	});
 
@@ -304,48 +123,4 @@ describe("TranscriptionService", () => {
 		});
 	});
 
-	describe("createChunkFiles small-file fast path (#168 / PR #204 review)", () => {
-		type ChunkFn = (args: {
-			buffer: ArrayBuffer;
-			basename: string;
-			extension: string;
-			mimeType: string;
-		}) => Promise<File[]>;
-
-		function createChunkFiles(): ChunkFn {
-			const service = new TranscriptionService(createMockPlugin());
-			return (service as unknown as { createChunkFiles: ChunkFn })
-				.createChunkFiles.bind(service);
-		}
-
-		test("sends a small m4a as a single original file instead of WAV-splitting it", async () => {
-			// A 1 KB m4a is well under the 20 MB chunk size, so it must not be
-			// converted to WAV (which would balloon it into many chunks and reset
-			// diarization speaker labels). It should be one intact .m4a file.
-			const files = await createChunkFiles()({
-				buffer: new ArrayBuffer(1024),
-				basename: "episode",
-				extension: "m4a",
-				mimeType: "audio/mp4",
-			});
-
-			expect(files).toHaveLength(1);
-			expect(files[0].name).toBe("episode.m4a");
-			expect(files[0].type).toBe("audio/mp4");
-			expect(files[0].name).not.toContain(".wav");
-		});
-
-		test("still sends a small mp3 as a single original file (unchanged)", async () => {
-			const files = await createChunkFiles()({
-				buffer: new ArrayBuffer(2048),
-				basename: "episode",
-				extension: "mp3",
-				mimeType: "audio/mpeg",
-			});
-
-			expect(files).toHaveLength(1);
-			expect(files[0].name).toBe("episode.mp3");
-			expect(files[0].size).toBe(2048);
-		});
-	});
 });
