+++ src/services/TranscriptionService.ts
@@ -6,6 +6,7 @@ import { TranscriptTemplateEngine } from "../TemplateEngine";
 import { ensureFolderExists } from "../utility/ensureFolderExists";
 import type { Episode } from "src/types/Episode";
 import { getEpisodeTranscriptPath } from "src/utility/getEpisodeTranscriptPath";
+import { createChunkFiles, getMimeType } from "./audioChunker";
 import {
 	type DiarizationAudio,
 	type DiarizationProviderId,
@@ -61,9 +62,6 @@ export class TranscriptionService {
 	private client: OpenAI | null = null;
 	private cachedApiKey: string | null = null;
 	private MAX_RETRIES = 3;
-	private readonly CHUNK_SIZE_BYTES = 20 * 1024 * 1024;
-	private readonly WAV_HEADER_SIZE = 44;
-	private readonly PCM_BYTES_PER_SAMPLE = 2;
 	private readonly MAX_CONCURRENT_TRANSCRIPTIONS = 2;
 	private readonly MAX_CONCURRENT_CHUNK_TRANSCRIPTIONS = 3;
 	private pendingEpisodes: Episode[] = [];
@@ -168,7 +166,7 @@ export class TranscriptionService {
 				extension: fileExtension,
 				basename,
 			} = await getEpisodeAudioBuffer(episode);
-			const mimeType = this.getMimeType(fileExtension);
+			const mimeType = getMimeType(fileExtension);
 
 			const transcriptBody = await this.buildTranscriptBody(
 				{
@@ -222,7 +220,7 @@ export class TranscriptionService {
 		}
 
 		updateNotice("Creating audio chunks...");
-		const files = await this.createChunkFiles(audio);
+		const files = await createChunkFiles(audio);
 		updateNotice("Starting transcription...");
 		const transcription = await this.transcribeChunks(files, updateNotice);
 		return transcription.replace(/\.\s+/g, ".\n\n");
@@ -252,242 +250,13 @@ export class TranscriptionService {
 		// OpenAI diarization shares Whisper's 25 MB/request cap, so reuse the same
 		// chunking. Speaker labels can differ across chunks on a long episode.
 		updateNotice("Creating audio chunks...");
-		const chunkFiles = await this.createChunkFiles(audio);
+		const chunkFiles = await createChunkFiles(audio);
 		return diarizeWithOpenAI({
 			getClient: () => this.getClient(),
 			chunkFiles,
 			maxRetries: this.MAX_RETRIES,
 			onProgress: updateNotice,
 		});
-	}
-
-	private async createChunkFiles({
-		buffer,
-		basename,
-		extension,
-		mimeType,
-	}: {
-		buffer: ArrayBuffer;
-		basename: string;
-		extension: string;
-		mimeType: string;
-	}): Promise<File[]> {
-		// A file that already fits in a single request needs neither conversion nor
-		// splitting: send the original (compressed) bytes. OpenAI accepts m4a/mp3/etc.
-		// directly under the upload limit, so this skips the m4a->WAV path below for
-		// small m4a episodes — that path only exists to SAFELY SPLIT an m4a (which
-		// can't be byte-split) and would otherwise balloon a small m4a into many
-		// uncompressed WAV chunks, multiplying requests and, for diarization,
-		// resetting speaker labels at artificial chunk boundaries (#168 / PR #204 review).
-		if (buffer.byteLength <= this.CHUNK_SIZE_BYTES) {
-			return [
-				new File([buffer], `${basename}.${extension}`, { type: mimeType }),
-			];
-		}
-
-		if (this.shouldConvertToWav(extension, mimeType)) {
-			const wavChunks = await this.convertToWavChunks(buffer, basename);
-			if (wavChunks.length > 0) {
-				return wavChunks;
-			}
-		}
-
-		return this.createBinaryChunkFiles(buffer, basename, extension, mimeType);
-	}
-
-	private shouldConvertToWav(extension: string, mimeType: string): boolean {
-		const normalizedExtension = extension.toLowerCase();
-		return normalizedExtension === "m4a" || mimeType === "audio/mp4";
-	}
-
-	private createBinaryChunkFiles(
-		buffer: ArrayBuffer,
-		basename: string,
-		extension: string,
-		mimeType: string,
-	): File[] {
-		if (buffer.byteLength <= this.CHUNK_SIZE_BYTES) {
-			return [
-				new File([buffer], `${basename}.${extension}`, {
-					type: mimeType,
-				}),
-			];
-		}
-
-		const files: File[] = [];
-		for (
-			let offset = 0, index = 0;
-			offset < buffer.byteLength;
-			offset += this.CHUNK_SIZE_BYTES, index++
-		) {
-			const chunk = buffer.slice(offset, offset + this.CHUNK_SIZE_BYTES);
-			files.push(
-				new File([chunk], `${basename}.part${index}.${extension}`, {
-					type: mimeType,
-				}),
-			);
-		}
-
-		return files;
-	}
-
-	private async convertToWavChunks(
-		buffer: ArrayBuffer,
-		basename: string,
-	): Promise<File[]> {
-		const audioContext = this.createAudioContext();
-		if (!audioContext) return [];
-
-		try {
-			const audioBuffer = await audioContext.decodeAudioData(buffer.slice(0));
-			return this.renderWavChunks(audioBuffer, basename);
-		} catch (error) {
-			console.warn("Failed to convert audio buffer for transcription", error);
-			return [];
-		} finally {
-			try {
-				await audioContext.close();
-			} catch (error) {
-				console.warn("Failed to close audio context", error);
-			}
-		}
-	}
-
-	private createAudioContext(): AudioContext | null {
-		if (typeof window === "undefined") {
-			return null;
-		}
-
-		const contextCtor =
-			window.AudioContext ||
-			(window as typeof window & { webkitAudioContext?: typeof AudioContext })
-				.webkitAudioContext;
-		if (!contextCtor) {
-			return null;
-		}
-
-		return new contextCtor();
-	}
-
-	private renderWavChunks(audioBuffer: AudioBuffer, basename: string): File[] {
-		const numChannels = audioBuffer.numberOfChannels;
-		const bytesPerFrame = numChannels * this.PCM_BYTES_PER_SAMPLE;
-		const availableBytesPerChunk = this.CHUNK_SIZE_BYTES - this.WAV_HEADER_SIZE;
-		const maxSamplesPerChunk = Math.max(
-			1,
-			Math.floor(availableBytesPerChunk / bytesPerFrame),
-		);
-		const channelData = Array.from({ length: numChannels }, (_, channelIndex) =>
-			audioBuffer.getChannelData(channelIndex),
-		);
-		const files: File[] = [];
-		let chunkIndex = 0;
-
-		for (
-			let startSample = 0;
-			startSample < audioBuffer.length;
-			startSample += maxSamplesPerChunk
-		) {
-			const endSample = Math.min(
-				audioBuffer.length,
-				startSample + maxSamplesPerChunk,
-			);
-			const wavBuffer = this.renderWavBuffer(
-				channelData,
-				audioBuffer.sampleRate,
-				startSample,
-				endSample,
-			);
-			files.push(
-				new File([wavBuffer], `${basename}.part${chunkIndex}.wav`, {
-					type: "audio/wav",
-				}),
-			);
-			chunkIndex++;
-		}
-
-		return files;
-	}
-
-	private renderWavBuffer(
-		channelData: Float32Array[],
-		sampleRate: number,
-		startSample: number,
-		endSample: number,
-	): ArrayBuffer {
-		const numChannels = channelData.length;
-		const sampleCount = Math.max(0, endSample - startSample);
-		const blockAlign = numChannels * this.PCM_BYTES_PER_SAMPLE;
-		const buffer = new ArrayBuffer(
-			this.WAV_HEADER_SIZE + sampleCount * blockAlign,
-		);
-		const view = new DataView(buffer);
-		this.writeWavHeader(view, sampleRate, numChannels, sampleCount);
-		let offset = this.WAV_HEADER_SIZE;
-
-		for (let i = 0; i < sampleCount; i++) {
-			for (let channel = 0; channel < numChannels; channel++) {
-				const sample = channelData[channel][startSample + i] ?? 0;
-				const clamped = Math.max(-1, Math.min(1, sample));
-				const intSample =
-					clamped < 0
-						? clamped * 0x8000
-						: clamped * 0x7fff;
-				view.setInt16(offset, Math.round(intSample), true);
-				offset += this.PCM_BYTES_PER_SAMPLE;
-			}
-		}
-
-		return buffer;
-	}
-
-	private writeWavHeader(
-		view: DataView,
-		sampleRate: number,
-		numChannels: number,
-		sampleCount: number,
-	): void {
-		const blockAlign = numChannels * this.PCM_BYTES_PER_SAMPLE;
-		const byteRate = sampleRate * blockAlign;
-		const dataSize = sampleCount * blockAlign;
-		this.writeString(view, 0, "RIFF");
-		view.setUint32(4, 36 + dataSize, true);
-		this.writeString(view, 8, "WAVE");
-		this.writeString(view, 12, "fmt ");
-		view.setUint32(16, 16, true);
-		view.setUint16(20, 1, true);
-		view.setUint16(22, numChannels, true);
-		view.setUint32(24, sampleRate, true);
-		view.setUint32(28, byteRate, true);
-		view.setUint16(32, blockAlign, true);
-		view.setUint16(34, this.PCM_BYTES_PER_SAMPLE * 8, true);
-		this.writeString(view, 36, "data");
-		view.setUint32(40, dataSize, true);
-	}
-
-	private writeString(view: DataView, offset: number, str: string): void {
-		for (let i = 0; i < str.length; i++) {
-			view.setUint8(offset + i, str.charCodeAt(i));
-		}
-	}
-
-	private getMimeType(fileExtension: string): string {
-		switch (fileExtension.toLowerCase()) {
-			case "mp3":
-				return "audio/mp3";
-			case "m4a":
-				return "audio/mp4";
-			case "ogg":
-				return "audio/ogg";
-			case "wav":
-				return "audio/wav";
-			case "flac":
-				return "audio/flac";
-			case "webm":
-				return "audio/webm";
-			default:
-				return "audio/mpeg";
-		}
 	}
 
 	private async transcribeChunks(
